make route stats context as pure class
