+++ envoy/router/context.h
@@ -35,5 +35,23 @@ public:
   virtual GenericConnPoolFactory& genericConnPoolFactory() PURE;
 };
 
+/**
+ * RouteStatsContext defines config needed to generate all route level stats.
+ */
+class RouteStatsContext {
+public:
+virtual ~RouteStatsContext() = default;
+
+  /**
+   * @return stat name for route specific stats.
+   */
+  virtual const Stats::StatName statName() const PURE;
+
+    /**
+   * @return a struct containing stats for route stats.
+   */
+  virtual const RouteStats& stats() const PURE;
+};
+
 } // namespace Router
 } // namespace Envoy
