+++ source/common/router/context_impl.h
@@ -53,5 +53,22 @@ private:
   GenericConnPoolFactory* generic_conn_pool_factory_;
 };
 
+class RouteStatsContextImpl: public RouteStatsContext {
+public:
+  explicit RouteStatsContextImpl(Stats::Scope& scope, const RouteStatNames& route_stat_names,
+                    const Stats::StatName& vhost_stat_name, const std::string& stat_prefix);
+
+~RouteStatsContextImpl() override = default;
+
+  const Stats::StatName statName() const override { return route_stat_name_; }
+  const RouteStats& stats() const override { return route_stats_; }
+
+private:
+  const Stats::StatNameManagedStorage route_stat_name_storage_;
+  Stats::ScopeSharedPtr route_stats_scope_;
+  Stats::StatName route_stat_name_;
+  RouteStats route_stats_;
+};
+
 } // namespace Router
 } // namespace Envoy
